+++ users-module/src/main/java/com/example/users/jpa/repository/ClienteRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.users.jpa.entity.Cliente;
 
-public interface ClienteRepository extends JpaRepository<Cliente, Integer> {
+public interface ClienteRepository extends JpaRepository<Cliente, Long> {
 
 	public Optional<Cliente> findByIdentificacion(Integer identificacion);
